refactor(type_checker): align tsoptions structure with typescript-go
- core/parsedoptions.rs ports core/parsedoptions.go: ParsedOptions
  { compiler_options, file_names, project_references }; file names are
  expanded eagerly at parse time (tsgo parseJsonConfigFileContentWorker)
  instead of walking the filesystem per file_names() call.
- ParsedCommandLine wraps ParsedOptions like tsgo's ParsedConfig and
  exposes tsgo's accessors (compiler_options, file_names,
  project_references).
- Enum string maps move to tsoptions/enummaps.rs, mirroring
  internal/tsoptions/enummaps.go (enums stay in core).
- Helper names follow their tsgo counterparts:
  get_extends_config_path(_or_array), get_file_names_from_config_specs,
  handle_option_config_dir_template_substitution,
  get_substituted_path_with_config_dir_template,
  convert_compiler_options_from_json_worker.
