+++ crates/oxc_type_checker/src/core/compileroptions.rs
@@ -15,7 +15,6 @@
 
 use std::path::{Path, PathBuf};
 
-use cow_utils::CowUtils;
 use indexmap::IndexMap;
 use rustc_hash::FxBuildHasher;
 
@@ -54,29 +53,6 @@ pub enum ModuleKind {
     Preserve,
 }
 
-impl ModuleKind {
-    /// tsgo `moduleOptionMap` (`internal/tsoptions/enummaps.go`), matched case-insensitively.
-    /// `"none"` and unknown values both yield `None` (see the type docs).
-    pub fn from_str_ignore_case(value: &str) -> Option<Self> {
-        match value.cow_to_ascii_lowercase().as_ref() {
-            "commonjs" => Some(Self::CommonJs),
-            "amd" => Some(Self::Amd),
-            "system" => Some(Self::System),
-            "umd" => Some(Self::Umd),
-            "es6" | "es2015" => Some(Self::Es2015),
-            "es2020" => Some(Self::Es2020),
-            "es2022" => Some(Self::Es2022),
-            "esnext" => Some(Self::EsNext),
-            "node16" => Some(Self::Node16),
-            "node18" => Some(Self::Node18),
-            "node20" => Some(Self::Node20),
-            "nodenext" => Some(Self::NodeNext),
-            "preserve" => Some(Self::Preserve),
-            _ => None,
-        }
-    }
-}
-
 /// tsgo `core.ModuleResolutionKind`. tsc's `moduleResolution` option.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum ModuleResolutionKind {
@@ -87,20 +63,6 @@ pub enum ModuleResolutionKind {
     Node16,
     NodeNext,
     Bundler,
-}
-
-impl ModuleResolutionKind {
-    /// tsgo `moduleResolutionOptionMap`, matched case-insensitively.
-    pub fn from_str_ignore_case(value: &str) -> Option<Self> {
-        match value.cow_to_ascii_lowercase().as_ref() {
-            "node16" => Some(Self::Node16),
-            "nodenext" => Some(Self::NodeNext),
-            "bundler" => Some(Self::Bundler),
-            "classic" => Some(Self::Classic),
-            "node" | "node10" => Some(Self::Node10),
-            _ => None,
-        }
-    }
 }
 
 /// tsgo `core.ScriptTarget`. Variants are declared in tsgo's numeric order for `PartialOrd`
@@ -128,26 +90,6 @@ pub enum ScriptTarget {
 impl ScriptTarget {
     /// tsgo `ScriptTargetLatestStandard`.
     pub const LATEST_STANDARD: Self = Self::Es2025;
-
-    /// tsgo `targetOptionMap`, matched case-insensitively.
-    pub fn from_str_ignore_case(value: &str) -> Option<Self> {
-        match value.cow_to_ascii_lowercase().as_ref() {
-            "es5" => Some(Self::Es5),
-            "es6" | "es2015" => Some(Self::Es2015),
-            "es2016" => Some(Self::Es2016),
-            "es2017" => Some(Self::Es2017),
-            "es2018" => Some(Self::Es2018),
-            "es2019" => Some(Self::Es2019),
-            "es2020" => Some(Self::Es2020),
-            "es2021" => Some(Self::Es2021),
-            "es2022" => Some(Self::Es2022),
-            "es2023" => Some(Self::Es2023),
-            "es2024" => Some(Self::Es2024),
-            "es2025" => Some(Self::Es2025),
-            "esnext" => Some(Self::EsNext),
-            _ => None,
-        }
-    }
 }
 
 /// tsgo `core.ModuleDetectionKind`. tsc's `moduleDetection` option.
@@ -156,18 +98,6 @@ pub enum ModuleDetectionKind {
     Auto,
     Legacy,
     Force,
-}
-
-impl ModuleDetectionKind {
-    /// tsgo `moduleDetectionOptionMap`, matched case-insensitively.
-    pub fn from_str_ignore_case(value: &str) -> Option<Self> {
-        match value.cow_to_ascii_lowercase().as_ref() {
-            "auto" => Some(Self::Auto),
-            "legacy" => Some(Self::Legacy),
-            "force" => Some(Self::Force),
-            _ => None,
-        }
-    }
 }
 
 /// tsgo `core.JsxEmit`. tsc's `jsx` option.
@@ -180,36 +110,11 @@ pub enum JsxEmit {
     ReactJsxDev,
 }
 
-impl JsxEmit {
-    /// tsgo `jsxOptionMap`, matched case-insensitively.
-    pub fn from_str_ignore_case(value: &str) -> Option<Self> {
-        match value.cow_to_ascii_lowercase().as_ref() {
-            "preserve" => Some(Self::Preserve),
-            "react-native" => Some(Self::ReactNative),
-            "react-jsx" => Some(Self::ReactJsx),
-            "react-jsxdev" => Some(Self::ReactJsxDev),
-            "react" => Some(Self::React),
-            _ => None,
-        }
-    }
-}
-
 /// tsgo `core.NewLineKind`. tsc's `newLine` option.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum NewLineKind {
     CarriageReturnLineFeed,
     LineFeed,
-}
-
-impl NewLineKind {
-    /// tsgo `newLineOptionMap`, matched case-insensitively.
-    pub fn from_str_ignore_case(value: &str) -> Option<Self> {
-        match value.cow_to_ascii_lowercase().as_ref() {
-            "crlf" => Some(Self::CarriageReturnLineFeed),
-            "lf" => Some(Self::LineFeed),
-            _ => None,
-        }
-    }
 }
 
 /// Invokes the callback macro with every user-settable compilerOptions field as
